+++ qelem/lib/common/qelem_theme.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-import 'AppPalette.dart';
+import 'app_palette.dart';
 
 class QelemTheme {
   static ThemeData qelemTheme = ThemeData(
